requirements.txt
anthropic
Pillow
pyyaml
openai-whisper
ffmpeg-python
yt-dlp
# streamlit  # Phase 4 UI用（オプション）
